Check if the fontawesome was previously downloaded before start a new download

+++ scripts/install/recipe/fonts.sh
@@ -12,9 +12,11 @@ mkdir -p $font_dir
 
 # Install single font from font-awesome
 #  This font file will be used for the i3 Bar
-cd $font_dir
-curl -o FontAwesome-webfont-latest.ttf https://github.com/FortAwesome/Font-Awesome/raw/master/fonts/fontawesome-webfont.ttf
-cp "$(dirname "$(readlink -fm "$0")")/fontawesome-webfont.ttf" $font_dir
+if [ ! -f "$font_dir/fontawesome-webfont.ttf" ]; then
+  cd $font_dir
+  curl -o FontAwesome-webfont-latest.ttf https://github.com/FortAwesome/Font-Awesome/raw/master/fonts/fontawesome-webfont.ttf
+  cp "$(dirname "$(readlink -fm "$0")")/fontawesome-webfont.ttf" $font_dir
+fi
 
 # Install powerline fonts
 mkdir $initialPath/customization
